✨ Refactor logging system by moving Logger and const definitions to App namespace and updating include paths
