+++ main.cpp
@@ -1,7 +1,7 @@
 # include <iostream>
 # include <string>
 # include <SFML/Graphics.hpp>
-# include "Logger.hpp"
+# include "App/Logger.hpp"
 # include "App/App.hpp"
