+++ src/App/DrawApp/DrawApp_Export.cpp
@@ -1,6 +1,6 @@
 #include <string>
 
-#include "Logger.hpp"
+#include "App/Logger.hpp"
 
 #include "App/DrawApp/DrawApp.hpp"
 #include "Exceptions/DrawAppExceptions.hpp"
